+++ src/main/java/com/ssy/dao/NoteEntityMapper.java
@@ -4,13 +4,13 @@ import com.ssy.entity.NoteEntity;
 import com.ssy.entity.NoteEntityWithBLOBs;
 
 public interface NoteEntityMapper {
-    int deleteByPrimaryKey(Long noteid);
+    int deleteByPrimaryKey(Integer noteid);
 
     int insert(NoteEntityWithBLOBs record);
 
     int insertSelective(NoteEntityWithBLOBs record);
 
-    NoteEntityWithBLOBs selectByPrimaryKey(Long noteid);
+    NoteEntityWithBLOBs selectByPrimaryKey(Integer noteid);
 
     int updateByPrimaryKeySelective(NoteEntityWithBLOBs record);
