+++ src/main/java/com/ssy/entity/CourseEntity.java
@@ -3,7 +3,7 @@ package com.ssy.entity;
 import java.util.Date;
 
 public class CourseEntity {
-    private Long courseid;
+    private Integer courseid;
 
     private String teacherid;
 
@@ -11,21 +11,21 @@ public class CourseEntity {
 
     private Date coursetime;
 
-    private Long score5;
+    private Integer score5;
 
-    private Long score4;
+    private Integer score4;
 
-    private Long score3;
+    private Integer score3;
 
-    private Long score2;
+    private Integer score2;
 
-    private Long score1;
+    private Integer score1;
 
-    public Long getCourseid() {
+    public Integer getCourseid() {
         return courseid;
     }
 
-    public void setCourseid(Long courseid) {
+    public void setCourseid(Integer courseid) {
         this.courseid = courseid;
     }
 
@@ -53,43 +53,43 @@ public class CourseEntity {
         this.coursetime = coursetime;
     }
 
-    public Long getScore5() {
+    public Integer getScore5() {
         return score5;
     }
 
-    public void setScore5(Long score5) {
+    public void setScore5(Integer score5) {
         this.score5 = score5;
     }
 
-    public Long getScore4() {
+    public Integer getScore4() {
         return score4;
     }
 
-    public void setScore4(Long score4) {
+    public void setScore4(Integer score4) {
         this.score4 = score4;
     }
 
-    public Long getScore3() {
+    public Integer getScore3() {
         return score3;
     }
 
-    public void setScore3(Long score3) {
+    public void setScore3(Integer score3) {
         this.score3 = score3;
     }
 
-    public Long getScore2() {
+    public Integer getScore2() {
         return score2;
     }
 
-    public void setScore2(Long score2) {
+    public void setScore2(Integer score2) {
         this.score2 = score2;
     }
 
-    public Long getScore1() {
+    public Integer getScore1() {
         return score1;
     }
 
-    public void setScore1(Long score1) {
+    public void setScore1(Integer score1) {
         this.score1 = score1;
     }
 }
